Refactor (amend): Apply rewrites asynchronously
This helps to remove the major IO bottleneck when applying rewrites;
multiple rewrite operations can happen in parallel, with the event
loop handling the sequencing when multiple tasks are waiting on the build of
an OID or commit. To avoid passing awaitables everywhere and to enforce
discipline around when operations are awaited, the main branch building code
retains a rewrite handle which is looked up to get an awaitable in the backend.